anomaly-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
httpx==0.28.1
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic==2.10.6